// generated/src/app/api-models/matrix-determ-y.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MatrixDetermY {
  matrixDeterminant: string;
  matrixSeq: number;
  yAxisDescription: string;
  familySizeFrom: number;
  familySizeThru: number;
  ageFrom: number;
  ageThru: number;
  gender: string;
  spouseFlag: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}